+++ CHANGELOG.md
@@ -52,6 +52,8 @@
 - Fixed a bug where version reported for the controller manager was missing
   due to incorrect linker flags in container builds.
   [#1943](https://github.com/Kong/kubernetes-ingress-controller/issues/1943)
+- The admission controller now also validates KongClusterPlugin resources.
+  [#1764](https://github.com/Kong/kubernetes-ingress-controller/issues/1764)
 
 ## [2.0.3] - 2021/10/19